pyproject.toml
[project]
name = "marchent-card-mastercard"
version = "0.1.0"
description = ""
authors = [
    {name = "Soumya2024",email = "habra4561@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
